Move curve operations into separate file
Better separate concerns per file so that draw.js is concerned only with
drawing points. Previously it dealt with anything that required a
semantic relationship between points (which excludes rotate points or
scale points because the points in those cases are not related).

Curves are embodied in those semantic relationships. For example, p1 and
p2 are related in that p1 is the starting point of a curve, p2 is the
ending point of a curve, and the control points (cp1, cp2) of the curve
are held by the ending point.