Generalize degree elevation and getting value
Previously we had getCubicControlPoints, which is essentially asking to
elevate degrees until cubic (and it was doing it incorrectly for linear
curves). Now we have elevateDegree which is generalized to any degree.

Since getting the value of linear, quadratic, and cubic curves are used
prominently and it needs to be a function to be passed to higher order
functions such as makePoint, they were made into functions.

The call signature for these functions will be standardized as
func(t, ...points)