Refactor curve interpolation
interpolateCurve now returns a list of points matching the query rather
than modifying the query. The filtering on t=[0,1] is now done
in one place (interpolateCurve) instead of the delegated functions. They
return the list of t (the curve parameter) that satisfy the
requirements.

Note the interpolation unit tests will occasionally fail because solving
the cubic equation can involve many steps through which precision is
lost.