Attach control points to p2 after splitting
Previously they were attached to the outer object (left or right), but
in most use cases we need to attach them to p2 immediately after calling
split anyway. This also encapsulates the degree of the curve from the
caller since they don't have to check whether cp1 or cp2 exists for
attachment.